fix(qt6): cast QChar::unicode() to ushort for QString::arg() (#15)
QChar::unicode() returns char16_t on Qt 6 (ushort on Qt 5), and
QString::arg() has no char16_t overload, so escapeString() failed to
compile with Qt 6 + clang:

    error: no matching member function for call to 'arg'

Both Linux Qt 6 builds pass because libstdc++/GCC selects the int
overload via integral promotion; libc++/clang treats the call as
ambiguous. Caught by the first native macOS build.

+++ Engine/NodeGroup.cpp
@@ -1655,7 +1655,9 @@ escapeString(const QString& str)
         } else {
 #if PY_MAJOR_VERSION >= 3
             // Python 3 strings are unicode
-            ret.append(QString::fromUtf8("\\u%1").arg(str[i].unicode(), 4, 16, QLatin1Char('0')));
+            // QChar::unicode() returns char16_t in Qt 6 (ushort in Qt 5), and
+            // QString::arg() has no char16_t overload.
+            ret.append(QString::fromUtf8("\\u%1").arg( static_cast<ushort>( str[i].unicode() ), 4, 16, QLatin1Char('0') ));
 #else
             // Python 2: convert to Utf8
             QByteArray utf8 = QString(str[i]).toUtf8();
